pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-japan-project"
version = "0.1.0"
description = "Local-first AI work platform demo for browser-based non-developer workflows."
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "streamlit>=1.30",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

